requirements.txt
fastapi
uvicorn
openai
httpx
pydantic
torch
vllm
PyYAML
